+++ src/main/scala/com/seanmcapp/util/parser/decoder/InstagramDecoder.scala
@@ -1,4 +1,4 @@
-package com.seanmcapp.util.parser
+package com.seanmcapp.util.parser.decoder
 
 case class InstagramAccountResponse(id: String)
