+++ src/main/scala/com/seanmcapp/util/parser/encoder/RouteEncoder.scala
@@ -1,3 +1,3 @@
-package com.seanmcapp.util.parser
+package com.seanmcapp.util.parser.encoder
 
 trait RouteEncoder extends TelegramOutputEncoder with DotaOutputEncoder
