+++ src/main/scala/com/seanmcapp/util/requestbuilder/TelegramRequestBuilder.scala
@@ -2,7 +2,8 @@ package com.seanmcapp.util.requestbuilder
 
 import com.seanmcapp.config.{StorageConf, TelegramConf}
 import com.seanmcapp.repository.instagram.Photo
-import com.seanmcapp.util.parser.{TelegramInputDecoder, TelegramOutputEncoder, TelegramResponse}
+import com.seanmcapp.util.parser.encoder.{TelegramOutputEncoder, TelegramResponse}
+import com.seanmcapp.util.parser.decoder.TelegramInputDecoder
 
 trait TelegramRequestBuilder extends TelegramInputDecoder with TelegramOutputEncoder {
